+++ changelog.md
@@ -30,6 +30,7 @@ Codion Change Log
 - Unit tests improved.
 - ProgressWorker.BuilderFactory added, replacing overloaded ProgressWorker.builder() methods.
 - FilterTableModel.Builder.Columns added, replacing the columns parameter in FilterTableModel.builder().
+- FilterTableModel.RowEditor renamed Editor.
 ### is.codion.swing.common.ui
 - Unit tests improved.
 - HintTextField.hintForegroundColor() bug fixed, returns null if either foreground or background is unspecified.